refactored the slider imgs and button part
