+++ server/index.js
@@ -10,7 +10,7 @@ app.use(express.static('client/dist'));
 
 console.log(1111)
 
-app.listen(3000,()=>{
+app.listen(4000,()=>{
   console.log('listening on port 3000');
 })
